novos arquivos, preparacao pra modularizacao
